pyproject.toml
[tool.poetry]
name = "spacetraders"
version = "0.11.0-rc18"
description = "Python API for Space Traders"
authors = ["Marco Ceppi <marco@ceppi.net>"]
license = "Apache-2.0"
readme = "README.md"
packages = [
    {include = "spacetraders_api_client"},
    {include = "spacetraders_generator"}
]

[tool.poetry.dependencies]
python = "^3.11"
httpx = "^0.24.0"
pydantic = "<3.0.0"
attrs = ">=21.3.0"
python-dateutil = "^2.8.2"

[tool.poetry.group.dev.dependencies]
black = "*"
isort = "*"
ruff = "*"
pytest = "^7.3.1"
openapi-python-client = "^0.17.1"
typer = {extras = ["all"], version = "^0.9.0"}

[tool.poetry.scripts]
spacetraders_generator = "spacetraders_generator.cli:app"

[tool.ruff]
# Allow lines to be as long as 120 characters.
line-length = 130

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
